embedded/tools/configurator/pyproject.toml
[project]
name = "feature-tree"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["mako==1.3.10", "oyaml==1.0"]

[tool.pyright]
reportAny = false
reportExplicitAny = false
reportImplicitStringConcatenation = false
reportUnusedCallResult = false
